# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: WP Maintenance Automation API
  description: |
    RESTful API for WordPress site maintenance automation including backups, upgrades, restores, and health checks.
    
    ## Authentication
    All endpoints except `/api/v1/health` and `/api/v1/auth/login` require authentication via Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## First Login
    At first use, call `/api/v1/auth/login` with a password that meets the requirements:
    - Minimum 8 characters
    - At least one uppercase letter
    - At least one lowercase letter
    - At least one digit
    
    This will create the internal admin user and return a 24-hour valid token.
  version: 1.0.0
  contact:
    name: WP Maintenance Automation
    url: https://github.com/andreacristalli/wp-maintenance-automation-go
servers:
  - url: https://localhost:8081
    description: Development server (HTTPS with self-signed cert)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      description: Single internal application profile used for web access and token ownership.
      properties:
        id:
          type: string
          description: User ID
        username:
          type: string
          description: Internal username
        display_name:
          type: string
          description: User-facing display name shown in the UI
        icon:
          type: string
          enum: [user, shield, bolt, globe, gear]
          description: Profile icon shown in the UI
        role:
          type: string
          enum: [admin, user, readonly]
          description: Stored role for the internal user, currently always admin in single-user mode
        force_pass:
          type: boolean
          description: Whether password change is required
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        display_name:
          type: string
        icon:
          type: string
          enum: [user, shield, bolt, globe, gear]
      required:
        - display_name
        - icon
    APIToken:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        token:
          type: string
          description: API token value (only returned on creation)
        name:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked:
          type: boolean
    Site:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        wp_ssh_host:
          type: string
        wp_ssh_port:
          type: integer
          default: 22
        wp_ssh_user:
          type: string
        wp_root:
          type: string
        db_host:
          type: string
        db_user:
          type: string
        db_password:
          type: string
        db_name:
          type: string
        restic_repository:
          type: string
        restic_password_file:
          type: string
        backup_dir:
          type: string
        retention_flags:
          type: string
        healthcheck_url:
          type: string
        staging_enabled:
          type: boolean
        staging_host:
          type: string
        staging_port:
          type: integer
        staging_user:
          type: string
        staging_root:
          type: string
    Backup:
      type: object
      properties:
        id:
          type: string
        site_id:
          type: string
        timestamp:
          type: string
        host:
          type: string
        wp_root:
          type: string
        wp_version:
          type: string
        db_name:
          type: string
        db_host:
          type: string
        db_version:
          type: string
        dump_file:
          type: string
        file_count:
          type: integer
        rs_exclude:
          type: string
        retention:
          type: string
        ssh_host:
          type: string
        ssh_user:
          type: string
        ssh_port:
          type: integer
        backup_size:
          type: integer
        checksum:
          type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
paths:
...
  /api/v1/auth/login:
    post:
      summary: Login / Create admin
      description: |
        At first login, creates the internal admin user with the provided password.
        Subsequent logins authenticate with the password and return a 24h token.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Password (min 8 chars, uppercase, lowercase, digit)
              required:
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user_id:
                        type: string
                      username:
                        type: string
                      role:
                        type: string
                      force_pass:
                        type: boolean
                      first_login:
                        type: boolean
                  success:
                    type: boolean
        '400':
          description: Invalid password or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid password
...
  /api/v1/users:
    get:
      summary: Get internal profile
      description: Returns the single internal user profile used by the application.
      operationId: getInternalProfile
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
                  success:
                    type: boolean
    put:
      summary: Update internal profile
      description: Updates the display name and icon for the single internal user.
      operationId: updateInternalProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated
        '400':
          description: Invalid request
...
  /api/v1/tokens:
    get:
      summary: List tokens
      operationId: listTokens
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIToken'
                  success:
                    type: boolean
    post:
      summary: Create token
      description: Creates a token for the single internal user. The owner cannot be selected.
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                duration:
                  type: integer
                  description: Duration in hours
      responses:
        '201':
          description: Token created
...
  /api/v1/sites:
    get:
      summary: List sites
      operationId: listSites
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
                  success:
                    type: boolean
    post:
      summary: Create site
      operationId: createSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Site'
      responses:
        '201':
          description: Site created
...
  /api/v1/backups:
    get:
      summary: List backups
      description: List backups, optionally filtered by site_id
      operationId: listBackups
      parameters:
        - name: site_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Backup'
                  success:
                    type: boolean
    post:
      summary: Create backup
      operationId: createBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                site_id:
                  type: string
              required:
                - site_id
      responses:
        '200':
          description: Backup initiated
...